Add the GUI that matches the mockup that Zak made.
This adds the four panels, one for the User Stories, one for the Users, one for the story summary, and one for the statistics. It also adds a button for starting the timer, as well as a drop down menu for the estimate options. It also addds a submit button.

Reported-by: Batyrlan Nurbekov, Sam Mailand, Zachary Zapatka, Miguel Mora