Add some interactivity to the GUI
This module will transfer the data that is in the drop down menu to the statistics box. It does not yet compute the statistics, but simply is to show that we can take the data and do something with it. It also prints "Timer Started" when the "Start Timer" button is pressed.