This specifically adds some interactivity to the GUI.
Change user story and user text box below header to tables, which does not change the functionality at this point.

Add responses so that the selected item in the drop down menu displays in the statistics box.

Add print statements to print out the selected item from the drop down menu.

Add print statments when the "Start Timer" button is pressed