Adds a basic GUI for when the user is either creating a new planning poker session, or a moderator is editing the game.
This basic GUI is currently not functional, but simply provides the basic layout that it will follow. This includes a place to type the name, a drop down menu for selecting a long or short term game (live or not), and selecting the deck of cards that will be used for the game.

There are two boxes that will hold the requirements that can be selected and one for the requirements that have already been selected.