Add a drop down menu for the date.
This now has a drop down menu for both month and day. Current the month is an enum. I added an action listener so that the number of days to select from change depending on the month. Currently it only works for either January or not January and the days are not the correct number.

This also now has a "Remove" button for the requirements.