austintrose.
Add the overview pane to the planning poker tab in WPI-Suit.

This includes a side panel that has all of the games, divided into 'answered, unanswered, and finished'

The right side of the screen is divided into four quadrants, one for picking which requirement the user is estimating, one for submitting an estimate, one for viewing other's votes (assuming the game has ended), and one for showing the statistics of the voting (assuming the game has ended).