Header: Adds the functionality of closing the existing new game tabs
Adds the functionality of closing the existing new game tabs. The class
that is responsible for this functionality was added -
ClosableTabComponent. This class contains two elements - button and title
of the tab. The title is assigned to the name of the tab specified during
the addition of the tab to the JTabbedPane.

Reported-by: bnurbekov
Signed-off-by: Batyrlan Nurbekov <bnurbekov@wpi.edu>