Moved Mailer from Janeway to PlanningPoker It's better for modularity and keeping our stuff within our own project.  And it actually works this way now.  No idea why it wouldn't work out of Janeway before.
The next step for the email is to actually set up a mock network and database to write test cases to make sure the user controllers work. Once that is done, then we hopefully we will have reorganized the game enough for me to figure out where we can safely put all out database-related objects.

Signed-out-by: Alec Thompson - ajthompson