Improv the way in which the requirements selected for the game are saved to the session.
Before this, the requirements were converted to a string, put into the JList, and then pulled out as a string. To save the list of requirements, there were then two for loops that searched through and looked at the name of each requirement and added it to the list of the requirements to be added to the game. This was incredibly inefficient, and has now been changed. The list of requirements in the JList are now stored as requirements, as JList automatically uses the toString method of an object when displaying information on the pane.

Names of the lists were also changed to make it more clear which list represented what.

Submitted by: sfmailand